fix EXPLAIN output
1. avoid ha_rows->double->ha_rows cast, for ~0ULL it behaves
   differently on Linux and Windows. This fixes the differences
   in derived_view and mdev13607 results between Linux and Windows.
2. EXPLAIN should show the number of rows as an unsigned number.

+++ sql/sql_explain.cc
@@ -508,7 +508,7 @@ int Explain_table_access::print_explain(select_result_sink *output, uint8 explai
   /* `rows` */
   if (rows_set)
   {
-    item_list.push_back(new Item_int((longlong) (ulonglong) rows, 
+    item_list.push_back(new Item_int((ulonglong) rows,
                          MY_INT64_NUM_DECIMAL_DIGITS));
   }
   else
